{{cookiecutter.project_repo_name}}/requirements.txt
oauth2client==3.0.0
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
numpy==1.20.2
pandas==1.2.3
streamlit==1.13.0
typedframe==0.7.1
cachetools==4.0.0
PyYAML==5.4.1